Move ID- and attribute- handling code into Element instead of Node
This also fixes a bug introduced in the comment-out commit, where internal code was relying on NodeImpl.prototype.id instead of using .getAttribute("id").
