+++ lib/jsdom/living/nodes/Element-impl.js
@@ -1,22 +1,26 @@
 "use strict";
-
+const vm = require("vm");
 const nwmatcher = require("nwmatcher/src/nwmatcher-noqsa");
 const idlUtils = require("../generated/utils");
 const NodeImpl = require("./Node-impl").implementation;
 const ParentNodeImpl = require("./ParentNode-impl").implementation;
 const attributes = require("../attributes");
+const namedPropertiesWindow = require("../named-properties-window");
 const NODE_TYPE = require("../node-type");
 const domToHtml = require("../../browser/domtohtml").domToHtml;
 const memoizeQuery = require("../../utils").memoizeQuery;
 const mapper = require("../../utils").mapper;
 const clone = require("../node").clone;
 const domSymbolTree = require("../helpers/internal-constants").domSymbolTree;
+const resetDOMTokenList = require("../dom-token-list").reset;
 const createLiveNodeList = require("../node-list").createLive;
 const DOMException = require("../../web-idl/DOMException");
 const createDOMTokenList = require("../dom-token-list").create;
 const attrGenerated = require("../generated/Attr");
+const MutationEvent = require("../generated/MutationEvent").interface;
 const validateNames = require("../helpers/validate-names");
 const listOfElementsWithClassNames = require("../node").listOfElementsWithClassNames;
+const proxiedWindowEventHandlers = require("../helpers/proxied-window-event-handlers");
 const NonDocumentTypeChildNode = require("./NonDocumentTypeChildNode-impl").implementation;
 
 // nwmatcher gets `document.documentElement` at creation-time, so we have to initialize lazily, since in the initial
@@ -55,6 +59,32 @@ function setInnerHTML(document, node, html) {
   }
 }
 
+function attachId(id, elm, doc) {
+  if (id && elm && doc) {
+    if (!doc._ids[id]) {
+      doc._ids[id] = [];
+    }
+    doc._ids[id].push(elm);
+  }
+}
+
+function detachId(id, elm, doc) {
+  if (id && elm && doc) {
+    if (doc._ids && doc._ids[id]) {
+      const elms = doc._ids[id];
+      for (let i = 0; i < elms.length; i++) {
+        if (elms[i] === elm) {
+          elms.splice(i, 1);
+          --i;
+        }
+      }
+      if (elms.length === 0) {
+        delete doc._ids[id];
+      }
+    }
+  }
+}
+
 class ElementImpl extends NodeImpl {
   constructor(args, privateData) {
     super(args, privateData);
@@ -67,6 +97,134 @@ class ElementImpl extends NodeImpl {
     this._prefix = null;
     this._localName = privateData.localName;
     this._attributes = attributes.createNamedNodeMap(this);
+  }
+
+  _attach() {
+    namedPropertiesWindow.nodeAttachedToDocument(this);
+
+    const id = this.getAttribute("id");
+    if (id) {
+      attachId(id, this, this._ownerDocument);
+    }
+
+    super._attach();
+  }
+
+  _detach() {
+    super._detach();
+
+    namedPropertiesWindow.nodeDetachedFromDocument(this);
+
+    const id = this.getAttribute("id");
+    if (id) {
+      detachId(id, this, this._ownerDocument);
+    }
+  }
+
+  _attrModified(name, value, oldValue) {
+    this._modified();
+    namedPropertiesWindow.elementAttributeModified(this, name, value, oldValue);
+
+    if (name === "id" && this._attached) {
+      const doc = this._ownerDocument;
+      detachId(oldValue, this, doc);
+      attachId(value, this, doc);
+    }
+
+    // TODO event handlers:
+    // The correct way to do this is lazy, and a bit more complicated; see
+    // https://html.spec.whatwg.org/multipage/webappapis.html#event-handler-content-attributes
+    // It would only be possible if we had proper getters/setters for every event handler, which we don't right now.
+    if (name.length > 2 && name[0] === "o" && name[1] === "n") {
+      if (value) {
+        const w = this._ownerDocument._global;
+        const self = proxiedWindowEventHandlers.has(name) && this._localName === "body" ? w : this;
+        const vmOptions = { filename: this._ownerDocument._URL, displayErrors: false };
+
+        // The handler code probably refers to functions declared globally on the window, so we need to run it in
+        // that context. In fact, it's worse; see
+        // https://code.google.com/p/chromium/codesearch#chromium/src/third_party/WebKit/Source/bindings/core/v8/V8LazyEventListener.cpp
+        // plus the spec, which show how multiple nested scopes are technically required. We won't implement that
+        // until someone asks for it, though.
+
+        // https://html.spec.whatwg.org/multipage/webappapis.html#the-event-handler-processing-algorithm
+
+        if (name === "onerror" && self === w) {
+          // https://html.spec.whatwg.org/multipage/webappapis.html#getting-the-current-value-of-the-event-handler
+          // step 10
+
+          self[name] = function (event, source, lineno, colno, error) {
+            w.__tempEventHandlerThis = this;
+            w.__tempEventHandlerEvent = event;
+            w.__tempEventHandlerSource = source;
+            w.__tempEventHandlerLineno = lineno;
+            w.__tempEventHandlerColno = colno;
+            w.__tempEventHandlerError = error;
+
+            try {
+              return vm.runInContext(`
+                (function (event, source, lineno, colno, error) {
+                  ${value}
+                }).call(__tempEventHandlerThis, __tempEventHandlerEvent, __tempEventHandlerSource,
+                        __tempEventHandlerLineno, __tempEventHandlerColno, __tempEventHandlerError)`, w, vmOptions);
+            } finally {
+              delete w.__tempEventHandlerThis;
+              delete w.__tempEventHandlerEvent;
+              delete w.__tempEventHandlerSource;
+              delete w.__tempEventHandlerLineno;
+              delete w.__tempEventHandlerColno;
+              delete w.__tempEventHandlerError;
+            }
+          };
+        } else {
+          self[name] = function (event) {
+            w.__tempEventHandlerThis = this;
+            w.__tempEventHandlerEvent = event;
+
+            try {
+              return vm.runInContext(`
+                (function (event) {
+                  ${value}
+                }).call(__tempEventHandlerThis, __tempEventHandlerEvent)`, w, vmOptions);
+            } finally {
+              delete w.__tempEventHandlerThis;
+              delete w.__tempEventHandlerEvent;
+            }
+          };
+        }
+      } else {
+        this[name] = null;
+      }
+    }
+
+    // TODO remove MutationEvents completely at some point
+    if (value !== oldValue && this._ownerDocument) {
+      const ev = this._ownerDocument.createEvent("MutationEvents");
+
+      let attrChange = MutationEvent.MODIFICATION;
+      if (value === null) {
+        attrChange = MutationEvent.REMOVAL;
+      }
+      if (oldValue === null) {
+        attrChange = MutationEvent.ADDITION;
+      }
+
+      ev.initMutationEvent("DOMAttrModified", true, false, this, oldValue, value, name, attrChange);
+      this.dispatchEvent(ev);
+    }
+
+    // update classList
+    if (name === "class" && value !== this.classList.toString()) {
+      resetDOMTokenList(this.classList, value);
+    }
+
+    if (value !== oldValue) {
+      const doc = this._ownerDocument;
+      const ev = doc.createEvent("MutationEvents");
+
+      ev.initMutationEvent("DOMSubtreeModified", true, false, this, null, null, null, null);
+      this.dispatchEvent(ev);
+    }
   }
 
   get namespaceURI() {
