+++ lib/jsdom/living/nodes/Node-impl.js
@@ -1,8 +1,6 @@
 "use strict";
-const vm = require("vm");
 
 const EventTargetImpl = require("../events/EventTarget-impl").implementation;
-const MutationEvent = require("../generated/MutationEvent").interface;
 const idlUtils = require("../generated/utils");
 
 const domSymbolTree = require("../helpers/internal-constants").domSymbolTree;
@@ -10,50 +8,15 @@ const simultaneousIterators = require("../../utils").simultaneousIterators;
 const DOMException = require("../../web-idl/DOMException");
 const NODE_TYPE = require("../node-type");
 const NODE_DOCUMENT_POSITION = require("../node-document-position");
-const resetDOMTokenList = require("../dom-token-list").reset;
 const createLiveNodeList = require("../node-list").createLive;
 const updateNodeList = require("../node-list").update;
 const updateHTMLCollection = require("../html-collection").update;
-const namedPropertiesWindow = require("../named-properties-window");
-const proxiedWindowEventHandlers = require("../helpers/proxied-window-event-handlers");
 const documentBaseURL = require("../helpers/document-base-url").documentBaseURL;
 const cloneNode = require("../node").clone;
 const attributes = require("../attributes");
 
-
-function attachId(id, elm, doc) {
-  if (id && elm && doc) {
-    if (!doc._ids[id]) {
-      doc._ids[id] = [];
-    }
-    doc._ids[id].push(elm);
-  }
-}
-
-function detachId(id, elm, doc) {
-  if (id && elm && doc) {
-    if (doc._ids && doc._ids[id]) {
-      const elms = doc._ids[id];
-      for (let i = 0; i < elms.length; i++) {
-        if (elms[i] === elm) {
-          elms.splice(i, 1);
-          --i;
-        }
-      }
-      if (elms.length === 0) {
-        delete doc._ids[id];
-      }
-    }
-  }
-}
-
 function getDocument(el) {
   return el.nodeType === NODE_TYPE.DOCUMENT_NODE ? el : el._ownerDocument;
-}
-
-function mutationEventsEnabled(el) {
-  return el.nodeType !== NODE_TYPE.ATTRIBUTE_NODE &&
-         getDocument(el).implementation._hasFeature("MutationEvents");
 }
 
 function isObsoleteNodeType(node) {
@@ -269,25 +232,23 @@ class NodeImpl extends EventTargetImpl {
       this._descendantAdded(this, newChildImpl);
     }
 
-    if (mutationEventsEnabled(this)) {
-      const doc = getDocument(this);
-      let ev = doc.createEvent("MutationEvents");
-      ev.initMutationEvent("DOMNodeInserted", true, false, this, null, null, null, null);
-      newChildImpl.dispatchEvent(ev);
+    const doc = getDocument(this);
+    let ev = doc.createEvent("MutationEvents");
+    ev.initMutationEvent("DOMNodeInserted", true, false, this, null, null, null, null);
+    newChildImpl.dispatchEvent(ev);
 
+    ev = doc.createEvent("MutationEvents");
+    ev.initMutationEvent("DOMSubtreeModified", true, false, this, null, null, null, null);
+    this.dispatchEvent(ev);
+
+    if (this.nodeType === NODE_TYPE.DOCUMENT_NODE || this._attachedToDocument) {
       ev = doc.createEvent("MutationEvents");
-      ev.initMutationEvent("DOMSubtreeModified", true, false, this, null, null, null, null);
-      this.dispatchEvent(ev);
+      ev.initMutationEvent("DOMNodeInsertedIntoDocument", false, false, null, null, null, null, null);
 
-      if (this.nodeType === NODE_TYPE.DOCUMENT_NODE || this._attachedToDocument) {
-        ev = doc.createEvent("MutationEvents");
-        ev.initMutationEvent("DOMNodeInsertedIntoDocument", false, false, null, null, null, null, null);
-
-        for (const el of domSymbolTree.treeIterator(newChildImpl)) {
-          if (el.nodeType === NODE_TYPE.ELEMENT_NODE) {
-            el.dispatchEvent(ev);
-            el._attachedToDocument = true;
-          }
+      for (const el of domSymbolTree.treeIterator(newChildImpl)) {
+        if (el.nodeType === NODE_TYPE.ELEMENT_NODE) {
+          el.dispatchEvent(ev);
+          el._attachedToDocument = true;
         }
       }
     }
@@ -332,113 +293,6 @@ class NodeImpl extends EventTargetImpl {
     }
   }
 
-  _attrModified(name, value, oldValue) {
-    this._modified();
-    namedPropertiesWindow.elementAttributeModified(this, name, value, oldValue);
-
-    if (name === "id" && this._attached) {
-      const doc = this._ownerDocument;
-      detachId(oldValue, this, doc);
-      attachId(value, this, doc);
-    }
-
-    // TODO event handlers:
-    // The correct way to do this is lazy, and a bit more complicated; see
-    // https://html.spec.whatwg.org/multipage/webappapis.html#event-handler-content-attributes
-    // It would only be possible if we had proper getters/setters for every event handler, which we don't right now.
-    if (name.length > 2 && name[0] === "o" && name[1] === "n") {
-      if (value) {
-        const w = this._ownerDocument._global;
-        const self = proxiedWindowEventHandlers.has(name) && this._localName === "body" ? w : this;
-        const vmOptions = { filename: this._ownerDocument._URL, displayErrors: false };
-
-        // The handler code probably refers to functions declared globally on the window, so we need to run it in
-        // that context. In fact, it's worse; see
-        // https://code.google.com/p/chromium/codesearch#chromium/src/third_party/WebKit/Source/bindings/core/v8/V8LazyEventListener.cpp
-        // plus the spec, which show how multiple nested scopes are technically required. We won't implement that
-        // until someone asks for it, though.
-
-        // https://html.spec.whatwg.org/multipage/webappapis.html#the-event-handler-processing-algorithm
-
-        if (name === "onerror" && self === w) {
-          // https://html.spec.whatwg.org/multipage/webappapis.html#getting-the-current-value-of-the-event-handler
-          // step 10
-
-          self[name] = function (event, source, lineno, colno, error) {
-            w.__tempEventHandlerThis = this;
-            w.__tempEventHandlerEvent = event;
-            w.__tempEventHandlerSource = source;
-            w.__tempEventHandlerLineno = lineno;
-            w.__tempEventHandlerColno = colno;
-            w.__tempEventHandlerError = error;
-
-            try {
-              return vm.runInContext(`
-                (function (event, source, lineno, colno, error) {
-                  ${value}
-                }).call(__tempEventHandlerThis, __tempEventHandlerEvent, __tempEventHandlerSource,
-                        __tempEventHandlerLineno, __tempEventHandlerColno, __tempEventHandlerError)`, w, vmOptions);
-            } finally {
-              delete w.__tempEventHandlerThis;
-              delete w.__tempEventHandlerEvent;
-              delete w.__tempEventHandlerSource;
-              delete w.__tempEventHandlerLineno;
-              delete w.__tempEventHandlerColno;
-              delete w.__tempEventHandlerError;
-            }
-          };
-        } else {
-          self[name] = function (event) {
-            w.__tempEventHandlerThis = this;
-            w.__tempEventHandlerEvent = event;
-
-            try {
-              return vm.runInContext(`
-                (function (event) {
-                  ${value}
-                }).call(__tempEventHandlerThis, __tempEventHandlerEvent)`, w, vmOptions);
-            } finally {
-              delete w.__tempEventHandlerThis;
-              delete w.__tempEventHandlerEvent;
-            }
-          };
-        }
-      } else {
-        this[name] = null;
-      }
-    }
-
-    // TODO remove MutationEvents completely at some point
-    if (value !== oldValue && this._ownerDocument &&
-      this._ownerDocument.implementation._hasFeature("MutationEvents")) {
-      const ev = this._ownerDocument.createEvent("MutationEvents");
-
-      let attrChange = MutationEvent.MODIFICATION;
-      if (value === null) {
-        attrChange = MutationEvent.REMOVAL;
-      }
-      if (oldValue === null) {
-        attrChange = MutationEvent.ADDITION;
-      }
-
-      ev.initMutationEvent("DOMAttrModified", true, false, this, oldValue, value, name, attrChange);
-      this.dispatchEvent(ev);
-    }
-
-    // update classList
-    if (name === "class" && value !== this.classList.toString()) {
-      resetDOMTokenList(this.classList, value);
-    }
-
-    if (mutationEventsEnabled(this) && value !== oldValue) {
-      const doc = getDocument(this);
-      const ev = doc.createEvent("MutationEvents");
-
-      ev.initMutationEvent("DOMSubtreeModified", true, false, this, null, null, null, null);
-      this.dispatchEvent(ev);
-    }
-  }
-
   replaceChild(node, child) {
     if (arguments.length < 2) {
       throw new TypeError("Not enough arguments to Node.prototype.replaceChild");
@@ -457,12 +311,6 @@ class NodeImpl extends EventTargetImpl {
 
   _attach() {
     this._attached = true;
-    namedPropertiesWindow.nodeAttachedToDocument(this);
-
-    const wrapper = idlUtils.wrapperForImpl(this);
-    if (wrapper.id) {
-      attachId(wrapper.id, this, this._ownerDocument);
-    }
 
     for (const child of domSymbolTree.childrenIterator(this)) {
       if (child._attach) {
@@ -473,12 +321,6 @@ class NodeImpl extends EventTargetImpl {
 
   _detach() {
     this._attached = false;
-
-    namedPropertiesWindow.nodeDetachedFromDocument(this);
-
-    if (this.id) {
-      detachId(this.id, this, this._ownerDocument);
-    }
 
     for (const child of domSymbolTree.childrenIterator(this)) {
       if (child._detach) {
@@ -496,24 +338,22 @@ class NodeImpl extends EventTargetImpl {
       throw new DOMException(DOMException.NOT_FOUND_ERR);
     }
 
-    if (mutationEventsEnabled(this)) {
-      const doc = getDocument(this);
-      let ev = doc.createEvent("MutationEvents");
+    const doc = getDocument(this);
+    let ev = doc.createEvent("MutationEvents");
 
-      ev.initMutationEvent("DOMNodeRemoved", true, false, this, null, null, null, null);
-      oldChildImpl.dispatchEvent(ev);
+    ev.initMutationEvent("DOMNodeRemoved", true, false, this, null, null, null, null);
+    oldChildImpl.dispatchEvent(ev);
 
-      ev = doc.createEvent("MutationEvents");
-      ev.initMutationEvent("DOMSubtreeModified", true, false, this, null, null, null, null);
-      this.dispatchEvent(ev);
+    ev = doc.createEvent("MutationEvents");
+    ev.initMutationEvent("DOMSubtreeModified", true, false, this, null, null, null, null);
+    this.dispatchEvent(ev);
 
-      ev = doc.createEvent("MutationEvents");
-      ev.initMutationEvent("DOMNodeRemovedFromDocument", false, false, null, null, null, null, null);
-      for (const el of domSymbolTree.treeIterator(oldChildImpl)) {
-        if (el.nodeType === NODE_TYPE.ELEMENT_NODE) {
-          el.dispatchEvent(ev);
-          el._attachedToDocument = false;
-        }
+    ev = doc.createEvent("MutationEvents");
+    ev.initMutationEvent("DOMNodeRemovedFromDocument", false, false, null, null, null, null, null);
+    for (const el of domSymbolTree.treeIterator(oldChildImpl)) {
+      if (el.nodeType === NODE_TYPE.ELEMENT_NODE) {
+        el.dispatchEvent(ev);
+        el._attachedToDocument = false;
       }
     }
